day2

+++ ft_bzero.c
@@ -10,9 +10,10 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_bzero(void *s, int n)
+//       void bzero(void *s, size_t n);
+void	ft_bzero(void *s, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (n > i)
