+++ ft_calloc.c
@@ -11,7 +11,8 @@
 /* ************************************************************************** */
 #include <stdlib.h>
 
-void	*ft_calloc(int nmemb, int size)
+//       void *calloc(size_t nmemb, size_t size);
+void	*ft_calloc(size_t nmemb, size_t size)
 {
 	void	*ptr;
